+++ raiden-dapp/CHANGELOG.md
@@ -4,8 +4,10 @@
 
 ### Added
 - [#691] Auto-suggesting hub on Goerli.
+- [#921] Transfer identifier on summary screen.
 
 [#691]: https://github.com/raiden-network/light-client/issues/691
+[#921]: https://github.com/raiden-network/light-client/issues/921
 
 ## [0.4.2] - 2020-03-05
 ### Fixed